+++ src/test/java/de/skuzzle/springboot/test/wiremock/TestStubCollectionAnnotation.java
@@ -6,6 +6,9 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
+import de.skuzzle.springboot.test.wiremock.stubs.HttpStub;
+import de.skuzzle.springboot.test.wiremock.stubs.Request;
+
 @Retention(RUNTIME)
 @Target({ ElementType.METHOD, ElementType.TYPE })
 @HttpStub(onRequest = @Request(toUrl = "/fromAnnotationCollection1"))