+++ src/test/java/de/skuzzle/springboot/test/wiremock/TestStubCollectionInterface.java
@@ -1,5 +1,8 @@
 package de.skuzzle.springboot.test.wiremock;
 
+import de.skuzzle.springboot.test.wiremock.stubs.HttpStub;
+import de.skuzzle.springboot.test.wiremock.stubs.Request;
+
 @HttpStub(onRequest = @Request(toUrl = "/fromInterfaceCollection1"))
 @HttpStub(onRequest = @Request(toUrl = "/fromInterfaceCollection2"))
 public interface TestStubCollectionInterface {